refactor: change totalitems from int to long (#11)
BREAKING CHANGE: Totalitems is a long instead of an int.

+++ src/Be.Vlaanderen.Basisregisters.Api/Search/Query.cs
@@ -31,7 +31,7 @@ namespace Be.Vlaanderen.Basisregisters.Api.Search
             FilteringHeader<TFilter> filtering,
             SortingHeader sorting,
             IPaginationRequest paginationRequest,
-            Func<IQueryable<T>, int> countFunc = null)
+            Func<IQueryable<T>, long> countFunc = null)
         {
             if (filtering == null)
                 throw new ArgumentNullException(nameof(filtering));
